refactor(frontend): decompose ChatInput.tsx (613 lines) into focused components
Extract three sub-components from the monolithic ChatInput:
- components/chat/ToolSelector.tsx — AI Chat / manual tool mode picker
- components/chat/FileAttachments.tsx — file badge display with upload status
- components/chat/PromptManager.tsx — load/save prompt buttons + modal

Also extract TOOL_CATEGORIES config to hooks/chat/tool-categories.ts
alongside the existing tool-labels.ts from PR #404.

ChatInput.tsx reduced from 613 to ~185 lines (textarea + submit + composition).
ChatPage now imports AI_CHAT_MODE from the shared config.
